+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: lmhlanga <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 12:43:48 by lmhlanga          #+#    #+#             */
-/*   Updated: 2019/05/28 11:11:23 by lmhlanga         ###   ########.fr       */
+/*   Updated: 2019/06/03 13:40:51 by lmhlanga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	char		*pd;
 	const char	*ps;
-	int			i;
+	size_t		i;
 	char		*temp;
 
 	pd = dest;
