+++ ft_strnequ.c
@@ -6,7 +6,7 @@
 /*   By: lmhlanga <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/25 14:17:28 by lmhlanga          #+#    #+#             */
-/*   Updated: 2019/05/27 13:52:10 by lmhlanga         ###   ########.fr       */
+/*   Updated: 2019/06/03 13:52:17 by lmhlanga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
